test(app): split titlebar rail tests out of session-tab-chip-contract
The right utility toggle clickability test and the tabs-slot full-height
test were sitting in session-tab-chip-contract.spec.ts because they
shared the "open 3+ tabs" fixture, but they assert a different contract:
the titlebar right rail's *layout* (toggle/tabs disjoint geometry, rail
full-height so the slot's border-l meets the panel separator). That has
nothing to do with chip visuals, × geometry, hover, or 4pt grid pinning.

Move both tests into their own spec so future titlebar-rail regressions
(e.g. toggle motion when the panel opens, the home/settings 0-width
gate, sidebar↔rail interactions) have an obvious home and the chip
spec stays focused on its single subject.

No behavior change — same 13 assertions across the two files.

Verified:
- bun run test:e2e session-tab-chip-contract.spec.ts (11/11)
- bun run test:e2e titlebar-right-rail-contract.spec.ts (2/2)

Refs PR #880.
