+++ packages/app/e2e/session/session-tab-chip-contract.spec.ts
@@ -401,72 +401,6 @@ test.describe("right-panel tab chip + × contract", () => {
     await page.getByRole("tab", REVIEW_TAB).hover() // also closes the hover test cleanly
   })
 
-  test("right utility toggle stays clickable with the full tab set open", async ({
-    page,
-    gotoSession,
-  }) => {
-    // Regression guard: PR #878 had the titlebar tabs slot absolute-positioned
-    // over `#pawwork-titlebar-right`, relying on `pointer-events-none` /
-    // `-auto` choreography to keep the toggle clickable through the overlay.
-    // That broke at 4+ tabs because the `+` button got pushed into the
-    // toggle's x range and `pointer-events: auto` (required for the `+`
-    // dropdown to open) intercepted clicks meant for the toggle.
-    // Fix: the tabs slot is now an in-flow flex sibling of the toggle inside
-    // the titlebar's right rail (see `Titlebar` comments) — disjoint
-    // geometry, no overlay, no click choreography. This test opens the full
-    // default-reachable tab set so any future regression that reintroduces
-    // overlap (e.g. re-absolutising the slot, growing toggle into the rail)
-    // fails fast.
-    await gotoSession()
-    await openRightPanel(page)
-    // Open the maximum default-reachable tab set so the strip is widest.
-    await page.locator("main").first().click()
-    await page.keyboard.press(`${modKey}+\\`) // fileTree.toggle → Files
-    await page.keyboard.press(`${modKey}+Shift+R`) // review.toggle → Review
-    await page.keyboard.press("Control+`") // terminal.toggle → Terminal (always Ctrl)
-    await page.mouse.move(0, 0)
-
-    const toggle = page.getByRole("button", { name: "Right utility panel" })
-    await expect(toggle).toHaveAttribute("aria-expanded", "true")
-    // If anything in the tabs-portal overlay swallows the click, this
-    // times out with "subtree intercepts pointer events".
-    await toggle.click({ timeout: 3_000 })
-    await expect(toggle).toHaveAttribute("aria-expanded", "false", { timeout: 2_000 })
-  })
-
-  test("tabs slot border-l spans the full titlebar height (no top/bottom seam break)", async ({
-    page,
-    gotoSession,
-  }) => {
-    // Regression guard from PR #880 followup review: with the tabs slot moved
-    // from absolute overlay to an in-flow flex sibling, its `self-stretch` only
-    // matches the parent flex container's content height. The titlebar root
-    // uses `items-center` (grid), which lets each grid cell collapse to its
-    // child's content box unless the cell opts out with `self-stretch` / `h-full`.
-    // If the right rail isn't full-height, the tabs slot's `border-l` also
-    // isn't full-height — it would visibly break above and below the toggle's
-    // 30px row, and stop meeting the right-panel body's `border-l` directly
-    // below the titlebar. This test pins the slot's painted height to the
-    // titlebar's own height so any future regression that drops the
-    // stretch chain fails fast.
-    await gotoSession()
-    await openRightPanel(page)
-    await page.mouse.move(0, 0)
-
-    const heights = await page.evaluate(() => {
-      const titlebar = document.querySelector('[data-component="titlebar-shell"]') as HTMLElement | null
-      const tabs = document.getElementById("pawwork-titlebar-tabs") as HTMLElement | null
-      return {
-        titlebar: titlebar ? Math.round(titlebar.getBoundingClientRect().height) : null,
-        tabs: tabs ? Math.round(tabs.getBoundingClientRect().height) : null,
-      }
-    })
-
-    expect(heights.titlebar).not.toBeNull()
-    expect(heights.tabs).not.toBeNull()
-    expect(heights.tabs).toBe(heights.titlebar)
-  })
-
   test("chip geometry pins to the 4pt grid (gap, padding, radius)", async ({ page, gotoSession }) => {
     // DESIGN.md L233 (4pt grid) + L305 (radius tiers sm/md/lg = 6/10/14).
     // The previous production values `gap-1.5` (6px) and `px-2.5` (10px)
